Docker/requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy<2.0
Pillow
scikit-learn
matplotlib
seaborn
pandas